requirements.txt
streamlit
azure-storage-blob==12.9.0
